+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.7.1
 
 - Prompt preview in test suite
+- "Latest" prompt version option for test suites
 
 v0.7.0 (Mahi-mahi)